[WIP] Refactoring rosie to create the core module
After talking to @cuducos, we started creating the core module that will make rosie generic.
By now what is done:
- Structure created
- Moved one classifier to the core module, ensuring that rosie keep working fine

What need to be done:
- Move core methods from __init__.py that is in chamber_of_deputies to __init__.py of core module
- Create `settings.py` with the definition of the classifiers used in chamber_of_deputies and `adapters.py` with the dataset_keys to chamber_of_deputies
- Prepare `rosie.py` to run each module separate (ex.: `python rosie.py run chamber_of_deputies`)
