+++ rosie/chamber_of_deputies/__init__.py
@@ -5,7 +5,7 @@ from sklearn.externals import joblib
 
 from rosie.chamber_of_deputies.dataset import Dataset
 from rosie.chamber_of_deputies.classifiers.election_expenses_classifier import ElectionExpensesClassifier
-from rosie.chamber_of_deputies.classifiers.invalid_cnpj_cpf_classifier import InvalidCnpjCpfClassifier
+from rosie.core.classifiers.invalid_cnpj_cpf_classifier import InvalidCnpjCpfClassifier
 from rosie.chamber_of_deputies.classifiers.meal_price_outlier_classifier import MealPriceOutlierClassifier
 from rosie.chamber_of_deputies.classifiers.monthly_subquota_limit_classifier import MonthlySubquotaLimitClassifier
 from rosie.chamber_of_deputies.classifiers.traveled_speeds_classifier import TraveledSpeedsClassifier
