+++ rosie/core/__init__.py
@@ -1,0 +1,12 @@
+import os.path
+
+import numpy as np
+from sklearn.externals import joblib
+
+class RosieCore:
+    """docstring for ClassName"""
+    def __init__(self, arg):
+        super(ClassName, self).__init__()
+        self.arg = arg
+
+
