+++ rosie/core/tests/fixtures/invalid_cnpj_cpf_classifier.csv
