+++ rosie/core/tests/test_invalid_cnpj_cpf_classifier.py
@@ -3,13 +3,13 @@ from unittest import TestCase
 import numpy as np
 import pandas as pd
 
-from rosie.chamber_of_deputies.classifiers.invalid_cnpj_cpf_classifier import InvalidCnpjCpfClassifier
+from rosie.core.classifiers.invalid_cnpj_cpf_classifier import InvalidCnpjCpfClassifier
 
 
 class TestInvalidCnpjCpfClassifier(TestCase):
 
     def setUp(self):
-        self.dataset = pd.read_csv('rosie/chamber_of_deputies/tests/fixtures/invalid_cnpj_cpf_classifier.csv',
+        self.dataset = pd.read_csv('rosie/core/tests/fixtures/invalid_cnpj_cpf_classifier.csv',
                                    dtype={'cnpj_cpf': np.str})
         self.subject = InvalidCnpjCpfClassifier()
